requirements.txt
asttokens==3.0.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
duckdb==1.4.3
executing==2.2.1
fonttools==4.61.1
hmmlearn==0.3.3
imbalanced-learn==0.14.1
imblearn==0.0
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.3
jupyter_client==8.7.0
jupyter_core==5.9.1
kiwisolver==1.4.9
matplotlib==3.10.8
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numpy==2.4.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
patsy==1.0.2
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
psutil==7.2.0
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.1.0
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.5
stack-data==0.6.3
statsmodels==0.14.6
threadpoolctl==3.6.0
tornado==6.5.4
traitlets==5.14.3
tzdata==2025.3
wcwidth==0.2.14
